# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "3.0.0",
    "title": "Handwriter Express Application",
    "description": "Handwriter Express Application Routes"
  },
  "host": "handwritingbackend.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/create": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "produces": [],
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User SignUP Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "Username - NoobMaster69"
                },
                "email": {
                  "type": "string",
                  "example": "XXXXXX@yyyy.com"
                },
                "password": {
                  "type": "string",
                  "example": "User-password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Create User in database with proper credentials.",
            "schema": {
              "type": "object",
              "properties": {
                "auth": {
                  "type": "boolean",
                  "example": true
                },
                "token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjYxYzZlMmZmMTYyMDUyZmNiYjI0OGZlYyIsInRva2VuX2lkIjoiODc5OEdUREQyODI2WkZQQiIsImlhdCI6MTY0MDQyNDE5Mn0.89JxJwWzT3DMSGCX3zQtPpf5rgSSJVS2cvpDwHUn-4U"
                },
                "message": {
                  "type": "string",
                  "example": "User created and Logged in"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Multiple 400 errors",
            "schema": {
              "type": "object",
              "properties": {
                "DetailsNotFound_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Username, Email, Password not found"
                    }
                  }
                },
                "UsernameNotFound_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Username not found"
                    }
                  }
                },
                "EmailNotFound_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Email not Found"
                    }
                  }
                },
                "PasswordNotFound_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Password not Found"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Multiple 401 errors",
            "schema": {
              "type": "object",
              "properties": {
                "UserEmailNotVerified_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Email not verified"
                    }
                  }
                },
                "OTPNotVerified_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "OTP not verified"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "409": {
            "description": "Multiple 409 errors",
            "schema": {
              "type": "object",
              "properties": {
                "UserEmail_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "User Email already found."
                    }
                  }
                },
                "UserName_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Username already used."
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error occured",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "produces": [],
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User Login Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "XXXXXX@yyyy.com"
                },
                "password": {
                  "type": "string",
                  "example": "User-password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User logged in",
            "schema": {
              "type": "object",
              "properties": {
                "auth": {
                  "type": "boolean",
                  "example": true
                },
                "token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjYxYzg5MGJiY2JmMjc1NDAwZDRkZGJmOSIsInRva2VuX2lkIjoiOTEwOVpGTVAxMDE0RUFBVSIsImlhdCI6MTY0MDUzNDI5N30.x92jESKmKLYFp1VbHBM9JMbUh8jfmR56ChrTIBb0_II"
                },
                "message": {
                  "type": "string",
                  "example": "User found and Logged in"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Multiple 400 errors",
            "schema": {
              "type": "object",
              "properties": {
                "DetailsNotFound_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Email, Password not found"
                    }
                  }
                },
                "EmailNotFound_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Email not Found"
                    }
                  }
                },
                "PasswordNotFound_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Password not Found"
                    }
                  }
                },
                "InvalidCredentials_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Password not Found"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error occured",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/logout": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Token BlackListed"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Forbidden"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/forget/password": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Registered Email Id",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "XXXXXX@yyyy.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User logged in",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Password reset OTP sent to user`s registered Email Id"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "User not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Email not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "User not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error occured"
          }
        }
      }
    },
    "/user/update/password": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Update User Password without logged in.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "resetPasswordToken": {
                  "type": "string",
                  "example": "AB12CD2"
                },
                "password": {
                  "type": "string",
                  "example": "XXXAAABB"
                },
                "email": {
                  "type": "string",
                  "example": "XXXXXX@yyyy.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Password reset OTP sent to user`s registered Email Id"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "User logged in"
          },
          "204": {
            "description": "User logged in",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Password reset successfull"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Multiple 400 Errors",
            "schema": {
              "type": "object",
              "properties": {
                "ResetPasswordTokenNotFound_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Reset Password Token not provided"
                    }
                  }
                },
                "PasswordNotFound_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Password not provided"
                    }
                  }
                },
                "OTPExpired_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "OTP Expired"
                    }
                  }
                },
                "EmailNotFound_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Email not Expired"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "User logged in",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No user found with this Email ID"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error occured"
          }
        }
      }
    },
    "/user/update/user/password": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Update User Password when logged in",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "XXXXXXewfefyyy"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User logged in",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Password update successfull"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "User not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bad Request"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Token BlackListed"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Forbidden"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error occured"
          }
        }
      }
    },
...
  },
...
}